AGENT/requirements.txt
openai>=1.12.0
langgraph>=0.2.0
langchain-core>=0.3.0
langsmith>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
